+++ src/test/java/nl/_42/restsecure/autoconfigure/test/MockedCrowdAuthenticationProviderConfig.java
@@ -4,7 +4,7 @@ import static java.util.Arrays.asList;
 
 import java.util.List;
 
-import nl._42.restsecure.autoconfigure.customizer.CustomAuthenticationProviders;
+import nl._42.restsecure.autoconfigure.CustomAuthenticationProviders;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;